[MI-3486] Update subscription monitor system (#294)
* [MI-3442] Optimise subscriptions monitor system (#261)

* [MI-3442] Optimise subscriptions monitor system

* [MI-3442] Optimise subscriptions system more by comparing with the subscriptions present on MS Teams

* [MI-3442] Refactor code more by using common function

* [MI-3442] Optimise chats subscription monitor

* [MI-3442] Fix and add new test cases and fix lint errors

* [MI-3442] Remove chat subscription logic from startSubscription function

* [MI-3442] Update logs, update store functions and fixed test cases

* [MI-3442] Add count of mock calls in the mock statements

* [MI-3442] Update go routine function in subscription monitor system

* [MI-3479] Fixed the issue of channel mentions not working from Teams to MM (#288)

Modified the struct for mention to include conversation ID as well
Modified the logic for mentioning everyone in a GM

* [MI-3478] Update monitor system for channel subscriptions (#289)

* [MI-3478] Update monitor system for channel subscriptions

* [MI-3478] Fix test cases

* [MI-3478] Fix test cases errors

* [MI-3486] Fix issue of creating map key using different ids

* [MI-3486] Review fixes

* [MI-3486] Review fixes: Modified some error messages in unit tests

* [MI-3486] Made a small change related to handling attachments greater than allowed size

---------

+++ server/monitor/subscriptions.go
@@ -1,98 +1,154 @@
 package monitor
 
 import (
+	"strings"
+	"sync"
 	"time"
 
+	"github.com/mattermost/mattermost-plugin-msteams-sync/server/msteams"
 	"github.com/mattermost/mattermost-plugin-msteams-sync/server/store/storemodels"
 )
 
-func (m *Monitor) checkChannelsSubscriptions() {
+func (m *Monitor) checkChannelsSubscriptions(msteamsSubscriptionsMap map[string]*msteams.Subscription) {
 	m.api.LogDebug("Checking for channels subscriptions")
-	subscriptions, err := m.store.ListChannelSubscriptionsToCheck()
+	links, err := m.store.ListChannelLinks()
 	if err != nil {
-		m.api.LogError("Unable to get the channel subscriptions", "error", err)
+		m.api.LogError("Unable to list channel links from DB", "error", err.Error())
 		return
 	}
-	m.api.LogDebug("Refreshing channels subscriptions", "count", len(subscriptions))
 
-	for _, subscription := range subscriptions {
-		link, _ := m.store.GetLinkByMSTeamsChannelID(subscription.TeamID, subscription.ChannelID)
-		if link == nil {
-			if err := m.store.DeleteSubscription(subscription.SubscriptionID); err != nil {
-				m.api.LogError("Unable to delete not-needed subscription", "error", err)
-			}
-			// Ignoring the error because can be the case that the subscription is no longer exists, in that case, it doesn't matter.
-			_ = m.client.DeleteSubscription(subscription.SubscriptionID)
-			continue
-		}
-
-		if time.Until(subscription.ExpiresOn) < (15 * time.Second) {
-			if err := m.recreateChannelSubscription(subscription.SubscriptionID, subscription.TeamID, subscription.ChannelID, subscription.Secret); err != nil {
-				m.api.LogError("Unable to recreate channel subscription properly", "error", err)
-			}
-		} else {
-			if err := m.refreshSubscription(subscription.SubscriptionID); err != nil {
-				m.api.LogDebug("Unable to refresh channel subscription properly", "error", err)
-				if err := m.recreateChannelSubscription(subscription.SubscriptionID, subscription.TeamID, subscription.ChannelID, subscription.Secret); err != nil {
-					m.api.LogError("Unable to recreate channel subscription properly", "error", err)
-				}
-			}
-		}
-	}
-}
-
-func (m *Monitor) checkChatsSubscriptions() {
-	m.api.LogDebug("Checking for chats subscriptions")
-	subscriptions, err := m.store.ListChatSubscriptionsToCheck()
+	subscriptions, err := m.store.ListChannelSubscriptions()
 	if err != nil {
-		m.api.LogError("Unable to get the chat subscriptions", "error", err)
+		m.api.LogError("Unable to get the channel subscriptions", "error", err.Error())
 		return
 	}
-	m.api.LogDebug("Refreshing chats subscriptions", "count", len(subscriptions))
 
+	channelSubscriptionsMap := make(map[string]*storemodels.ChannelSubscription)
 	for _, subscription := range subscriptions {
-		if time.Until(subscription.ExpiresOn) < (15 * time.Second) {
-			if err := m.recreateChatSubscription(subscription.SubscriptionID, subscription.UserID, subscription.Secret); err != nil {
-				m.api.LogError("Unable to recreate chat subscription properly", "error", err)
-			}
-		} else {
-			if err := m.refreshSubscription(subscription.SubscriptionID); err != nil {
-				m.api.LogDebug("Unable to refresh chat subscription properly", "error", err)
-				if err := m.recreateChatSubscription(subscription.SubscriptionID, subscription.UserID, subscription.Secret); err != nil {
-					m.api.LogError("Unable to recreate chat subscription properly", "error", err)
-				}
-			}
-		}
+		channelSubscriptionsMap[subscription.TeamID+subscription.ChannelID] = subscription
 	}
+
+	wg := sync.WaitGroup{}
+	ws := make(chan struct{}, 20)
+
+	for _, link := range links {
+		ws <- struct{}{}
+		wg.Add(1)
+		go func(link storemodels.ChannelLink) {
+			defer wg.Done()
+			mmSubscription, mmSubscriptionFound := channelSubscriptionsMap[link.MSTeamsTeam+link.MSTeamsChannel]
+			// Check if channel subscription is present for a link on Mattermost
+			if mmSubscriptionFound {
+				// Check if channel subscription is not present on MS Teams
+				if _, msteamsSubscriptionFound := msteamsSubscriptionsMap[mmSubscription.SubscriptionID]; !msteamsSubscriptionFound {
+					// Create channel subscription for the linked channel
+					m.recreateChannelSubscription(mmSubscription.SubscriptionID, mmSubscription.TeamID, mmSubscription.ChannelID, m.webhookSecret, false)
+					<-ws
+					return
+				} else if time.Until(mmSubscription.ExpiresOn) < (5 * time.Minute) {
+					if err := m.refreshSubscription(mmSubscription.SubscriptionID); err != nil {
+						m.api.LogDebug("Unable to refresh channel subscription", "error", err.Error())
+						m.recreateChannelSubscription(mmSubscription.SubscriptionID, mmSubscription.TeamID, mmSubscription.ChannelID, mmSubscription.Secret, true)
+					}
+				}
+			} else {
+				// Create channel subscription for the linked channel
+				m.recreateChannelSubscription("", link.MSTeamsTeam, link.MSTeamsChannel, m.webhookSecret, false)
+				<-ws
+				return
+			}
+			<-ws
+		}(link)
+	}
+	wg.Wait()
 }
 
-func (m *Monitor) checkGlobalSubscriptions() {
+// Commenting the below function as we are not creating any user type subscriptions
+// func (m *Monitor) checkChatsSubscriptions() {
+// 	m.api.LogDebug("Checking for chats subscriptions")
+// 	subscriptions, err := m.store.ListChatSubscriptionsToCheck()
+// 	if err != nil {
+// 		m.api.LogError("Unable to get the chat subscriptions", "error", err)
+// 		return
+// 	}
+// 	m.api.LogDebug("Refreshing chats subscriptions", "count", len(subscriptions))
+
+// 	for _, subscription := range subscriptions {
+// 		if time.Until(subscription.ExpiresOn) < (15 * time.Second) {
+// 			if err := m.recreateChatSubscription(subscription.SubscriptionID, subscription.UserID, subscription.Secret); err != nil {
+// 				m.api.LogError("Unable to recreate chat subscription properly", "error", err)
+// 			}
+// 		} else {
+// 			if err := m.refreshSubscription(subscription.SubscriptionID); err != nil {
+// 				m.api.LogDebug("Unable to refresh chat subscription properly", "error", err)
+// 				if err := m.recreateChatSubscription(subscription.SubscriptionID, subscription.UserID, subscription.Secret); err != nil {
+// 					m.api.LogError("Unable to recreate chat subscription properly", "error", err)
+// 				}
+// 			}
+// 		}
+// 	}
+// }
+
+func (m *Monitor) checkGlobalSubscriptions(msteamsSubscriptionsMap map[string]*msteams.Subscription, allChatsSubscription *msteams.Subscription) {
 	m.api.LogDebug("Checking for global subscriptions")
-	subscriptions, err := m.store.ListGlobalSubscriptionsToCheck()
+	subscriptions, err := m.store.ListGlobalSubscriptions()
 	if err != nil {
-		m.api.LogError("Unable to get the global subscriptions", "error", err)
+		m.api.LogError("Unable to get the chat subscriptions from store", "error", err.Error())
 		return
 	}
-	m.api.LogDebug("Refreshing global subscriptions", "count", len(subscriptions))
-	for _, subscription := range subscriptions {
-		if time.Until(subscription.ExpiresOn) < (15 * time.Second) {
-			if err := m.recreateGlobalSubscription(subscription.SubscriptionID, subscription.Secret); err != nil {
-				m.api.LogError("Unable to recreate global subscription properly", "error", err)
-			}
+
+	if len(subscriptions) == 0 {
+		if allChatsSubscription == nil {
+			m.CreateAndSaveChatSubscription(nil)
 		} else {
-			if err := m.refreshSubscription(subscription.SubscriptionID); err != nil {
-				m.api.LogDebug("Unable to refresh global subscription properly", "error", err)
-				if err := m.recreateGlobalSubscription(subscription.SubscriptionID, subscription.Secret); err != nil {
-					m.api.LogError("Unable to recreate global subscription properly", "error", err)
-				}
+			if err := m.store.SaveGlobalSubscription(storemodels.GlobalSubscription{SubscriptionID: allChatsSubscription.ID, Type: "allChats", ExpiresOn: allChatsSubscription.ExpiresOn, Secret: m.webhookSecret}); err != nil {
+				m.api.LogError("Unable to store all chats subscription in store", "subscriptionID", allChatsSubscription.ID, "error", err.Error())
 			}
 		}
+
+		return
+	}
+
+	mmSubscription := subscriptions[0]
+	// Check if all chats subscription is not present on MS Teams
+	if _, msteamsSubscriptionFound := msteamsSubscriptionsMap[mmSubscription.SubscriptionID]; !msteamsSubscriptionFound {
+		// Create all chats subscription on MS Teams
+		m.CreateAndSaveChatSubscription(mmSubscription)
+		return
+	}
+
+	if time.Until(mmSubscription.ExpiresOn) < (5 * time.Minute) {
+		if err := m.refreshSubscription(mmSubscription.SubscriptionID); err != nil {
+			m.api.LogDebug("Unable to refresh all chats subscription", "error", err.Error())
+			if err := m.recreateGlobalSubscription(mmSubscription.SubscriptionID, mmSubscription.Secret); err != nil {
+				m.api.LogError("Unable to recreate all chats subscription", "error", err.Error())
+			}
+		}
+	}
+}
+
+func (m *Monitor) CreateAndSaveChatSubscription(mmSubscription *storemodels.GlobalSubscription) {
+	newSubscription, err := m.client.SubscribeToChats(m.baseURL, m.webhookSecret, !m.useEvaluationAPI)
+	if err != nil {
+		m.api.LogError("Unable to create subscription for all chats", "error", err.Error())
+		return
+	}
+
+	if mmSubscription != nil {
+		if err := m.store.DeleteSubscription(mmSubscription.SubscriptionID); err != nil {
+			m.api.LogError("Unable to delete the old all chats subscription", "error", err.Error())
+		}
+	}
+
+	if err := m.store.SaveGlobalSubscription(storemodels.GlobalSubscription{SubscriptionID: newSubscription.ID, Type: "allChats", Secret: m.webhookSecret, ExpiresOn: newSubscription.ExpiresOn}); err != nil {
+		m.api.LogError("Unable to create subscription for all chats", "error", err.Error())
+		return
 	}
 }
 
 func (m *Monitor) recreateChatSubscription(subscriptionID, userID, secret string) error {
 	if err := m.client.DeleteSubscription(subscriptionID); err != nil {
-		m.api.LogDebug("Unable to delete old subscription, maybe it doesn't exist anymore in the server", "error", err)
+		m.api.LogDebug("Unable to delete old subscription, maybe it doesn't exist anymore in the server", "error", err.Error())
 	}
 
 	newSubscription, err := m.client.SubscribeToUserChats(userID, m.baseURL, m.webhookSecret, !m.useEvaluationAPI)
@@ -103,25 +159,34 @@ func (m *Monitor) recreateChatSubscription(subscriptionID, userID, secret string
 	return m.store.SaveChatSubscription(storemodels.ChatSubscription{SubscriptionID: newSubscription.ID, UserID: userID, Secret: secret, ExpiresOn: newSubscription.ExpiresOn})
 }
 
-func (m *Monitor) recreateChannelSubscription(subscriptionID, teamID, channelID, secret string) error {
-	if err := m.client.DeleteSubscription(subscriptionID); err != nil {
-		m.api.LogDebug("Unable to delete old subscription, maybe it doesn't exist anymore in the server", "error", err)
+func (m *Monitor) recreateChannelSubscription(subscriptionID, teamID, channelID, secret string, deleteFromClient bool) {
+	if deleteFromClient && subscriptionID != "" {
+		if err := m.client.DeleteSubscription(subscriptionID); err != nil {
+			m.api.LogDebug("Unable to delete old subscription, maybe it doesn't exist anymore in the server", "error", err.Error())
+		}
 	}
 
 	newSubscription, err := m.client.SubscribeToChannel(teamID, channelID, m.baseURL, m.webhookSecret)
 	if err != nil {
-		return err
+		m.api.LogError("Unable to create new subscription for the channel", "channelID", channelID, "error", err.Error())
+		return
 	}
 
-	if err = m.store.DeleteSubscription(subscriptionID); err != nil {
-		m.api.LogDebug("Unable to delete old channel subscription from DB", "subscriptionID", subscriptionID, "error", err.Error())
+	if subscriptionID != "" {
+		if err = m.store.DeleteSubscription(subscriptionID); err != nil {
+			m.api.LogDebug("Unable to delete old channel subscription from DB", "subscriptionID", subscriptionID, "error", err.Error())
+		}
 	}
-	return m.store.SaveChannelSubscription(storemodels.ChannelSubscription{SubscriptionID: newSubscription.ID, TeamID: teamID, ChannelID: channelID, Secret: secret, ExpiresOn: newSubscription.ExpiresOn})
+
+	if err := m.store.SaveChannelSubscription(storemodels.ChannelSubscription{SubscriptionID: newSubscription.ID, TeamID: teamID, ChannelID: channelID, Secret: secret, ExpiresOn: newSubscription.ExpiresOn}); err != nil {
+		m.api.LogError("Unable to store new subscription in DB", "subscriptionID", newSubscription.ID, "error", err.Error())
+		return
+	}
 }
 
 func (m *Monitor) recreateGlobalSubscription(subscriptionID, secret string) error {
 	if err := m.client.DeleteSubscription(subscriptionID); err != nil {
-		m.api.LogDebug("Unable to delete old subscription, maybe it doesn't exist anymore in the server", "error", err)
+		m.api.LogDebug("Unable to delete old subscription, maybe it doesn't exist anymore in the server", "error", err.Error())
 	}
 
 	newSubscription, err := m.client.SubscribeToChats(m.baseURL, secret, !m.useEvaluationAPI)
@@ -141,4 +206,24 @@ func (m *Monitor) refreshSubscription(subscriptionID string) error {
 		return err
 	}
 	return m.store.UpdateSubscriptionExpiresOn(subscriptionID, *newSubscriptionTime)
+}
+
+func (m *Monitor) GetMSTeamsSubscriptionsMap() (msteamsSubscriptionsMap map[string]*msteams.Subscription, allChatsSubscription *msteams.Subscription, err error) {
+	msteamsSubscriptions, err := m.client.ListSubscriptions()
+	if err != nil {
+		m.api.LogError("Unable to list MS Teams subscriptions", "error", err.Error())
+		return nil, nil, err
+	}
+
+	msteamsSubscriptionsMap = make(map[string]*msteams.Subscription)
+	for _, msteamsSubscription := range msteamsSubscriptions {
+		if strings.HasPrefix(msteamsSubscription.NotificationURL, m.baseURL) {
+			msteamsSubscriptionsMap[msteamsSubscription.ID] = msteamsSubscription
+			if strings.Contains(msteamsSubscription.Resource, "chats/getAllMessages") {
+				allChatsSubscription = msteamsSubscription
+			}
+		}
+	}
+
+	return
 }
